Started to working on directory structure improvement
