+++ frontend/src/components/CraftingPanel/CraftingPanel.js
@@ -1,12 +1,10 @@
 import React, { useState, useContext } from "react";
 import './craftingPanel.css';
-import { InputText } from "primereact/inputtext";
 import { Button } from "primereact/button";
 import DropdownMenu from "../OptionItemsMenu/DropdownMenu";
 import { ButtonContext } from "../../context/ButtonContext";
-import { Divider } from 'primereact/divider';
-import { Slider } from "primereact/slider";
-import { InputNumber } from 'primereact/inputnumber';
+import { TextHandler } from "./craftingComponents/TextHandler/TextHandler";
+import { StylingHandler } from "./craftingComponents/StylingHandler/StylingHandler";
  
 
 export function CraftingPanel() {
@@ -28,13 +26,6 @@ export function CraftingPanel() {
   const handleTextColorChange = (e) => {
     updateButtonTextSettings('textColor', e.target.value);
   };
-  const handleButtoTextnChange = (e) => {
-    updateButtonTextSettings('buttonText', e.target.value);
-  };
-
-  const handleBorderRadiusChange = (e) => {
-    updateButtonGeneralSettings('borderRadius', e.value);
-  }
 
   const handleDirectionUp = () => {
     updateHoverEffectSettings('hoverEffectDirection', 'translateY(-100%)');
@@ -80,38 +71,22 @@ export function CraftingPanel() {
       <div className="tab-content">
         {activeTab === 'General' && (
           <div className="settings-options">
-            <div className="settings-option">
-                {/* <div className="section-divider">Button Text</div> */}
-                {/* <Divider align="left">
-                  <div className="inline-flex align-items-center">
-                      <b>Text</b>
-                  </div>
-                </Divider> */}
-                <label htmlFor="ButtonText">Text</label>
-                <InputText variant="filled" id="ButtonText" placeholder="Craft" value={buttonTextSettings.buttonText} className="p-inputtext-sm" onChange={handleButtoTextnChange} />
-            </div>
-            <div className="two-options-container">
-              <div className="big-option-wrapper">
-                <Slider min='0' max='30' value={buttonGeneralSettings.borderRadius} onChange={handleBorderRadiusChange} className="w-full" />
-              </div>
-              <div className="small-option-wrapper" style={{ padding: '10px' }}>
-                <InputNumber inputStyle={{ width: '80px' }} maxLength={3} max={30} inputId="percent" value={buttonGeneralSettings.borderRadius} onValueChange={handleBorderRadiusChange} prefix="%" />
-              </div>
-            </div>
+            <TextHandler />
+            <StylingHandler />
           </div>
         )}
         {activeTab === 'Colors' && (
           <div className="settings-options">
-            <div className="settings-option">
-              {/* <label htmlFor="TextColor">Primary Color</label> */}
+            
               <div>Text Color</div>
               <input 
                 type="color"
                 id="TextColor" 
                 value={buttonTextSettings.textColor}
                 onChange={handleTextColorChange}
+                className="color-input"
                 />
-            </div>
+            
             <div className="settings-option">
               {/* <label htmlFor="bgColor">Secondary Color</label> */}
               <div>Background Color</div>
